Tekstit ym.

+++ src/CourseFinder.js
@@ -1,16 +1,15 @@
 import { useAuthenticator } from '@aws-amplify/ui-react';
 import React, { useState, useEffect, useCallback } from 'react';
 
-import { Alert, Button, TextField } from '@aws-amplify/ui-react';
+import { Alert, Button, TextField, Expander, ExpanderItem } from '@aws-amplify/ui-react';
 import Subscriptions from './Subscriptions';
 
-import {
-  Table,
-  TableCell,
-  TableBody,
-  TableHead,
-  TableRow,
-} from '@aws-amplify/ui-react';
+import CourseList from './CourseList';
+
+import Swal from 'sweetalert2'
+import withReactContent from 'sweetalert2-react-content'
+
+const MySwal = withReactContent(Swal)
 
 const CourseFinder = () => {
   const { user } = useAuthenticator((context) => [context.user]);
@@ -27,30 +26,8 @@ const CourseFinder = () => {
     setNotKeywords(event.target.value);
   }
 
-  const handleSubmit = (event) => {
-    setIsLoading(true);
-    setItems([]);
-    fetch(`${process.env.REACT_APP_API_URL_BASE}/courses?`
-      + new URLSearchParams({
-        keywords: keywords,
-        not_keywords: notKeywords,
-    }),{
-      headers: {Authorization: `Bearer ${user.signInUserSession.idToken.jwtToken}`}
-    })
-      .then(res => res.json())
-      .then(
-        (result) => {
-          setIsLoading(false);
-          setItems(result.courses);
-        },
-        (error) => {
-          setIsLoading(false);
-          setError(error);
-        }
-      )
-    event.preventDefault();
-  }
   const handleAddSubscription = (event) => {
+    setItems([]);
     setIsLoading(true);
     fetch(`${process.env.REACT_APP_API_URL_BASE}/subscriptions`,{
       method: "POST",
@@ -74,16 +51,50 @@ const CourseFinder = () => {
     event.preventDefault();
   }
   const handleDropSubscription = (subscriptionId) => {
+    setItems([]);
+    MySwal.fire({
+      title: <p>Oletko varma</p>,
+      showDenyButton: true,
+      showCancelButton: true,
+      confirmButtonText: 'Kyllä, poista',
+      denyButtonText: 'Ei, älä poista',
+    }).then((resp) => {
+      if (resp.isConfirmed) {
+        setIsLoading(true);
+        fetch(`${process.env.REACT_APP_API_URL_BASE}/subscriptions/${subscriptionId}`,{
+          method: "DELETE",
+          headers: {Authorization: `Bearer ${user.signInUserSession.idToken.jwtToken}`}
+        })
+          .then(res => res.json())
+          .then(
+            (result) => {
+              setIsLoading(false);
+              refreshSubscriptions();
+            },
+            (error) => {
+              setIsLoading(false);
+              setError(error);
+            }
+          )
+      }
+    })
+  }
+
+  const onShowCourses = (subscriptionId, keywords, notKeywords) => {
     setIsLoading(true);
-    fetch(`${process.env.REACT_APP_API_URL_BASE}/subscriptions/${subscriptionId}`,{
-      method: "DELETE",
+    setItems([]);
+    fetch(`${process.env.REACT_APP_API_URL_BASE}/courses?`
+      + new URLSearchParams({
+        keywords: keywords || '',
+        not_keywords: notKeywords || '',
+    }),{
       headers: {Authorization: `Bearer ${user.signInUserSession.idToken.jwtToken}`}
     })
       .then(res => res.json())
       .then(
         (result) => {
           setIsLoading(false);
-          refreshSubscriptions();
+          setItems(result.courses);
         },
         (error) => {
           setIsLoading(false);
@@ -120,56 +131,37 @@ const CourseFinder = () => {
   return (
     <>
 
-      {errorState ?? <Alert variation="error">{errorState}</Alert> }
+      {errorState && <Alert variation="error">{errorState}</Alert> }
 
       <Subscriptions
         subscriptions={subscriptions}
         onDropSubscription={handleDropSubscription}
+        onShowCourses={onShowCourses}
       />
 
-      <form onSubmit={handleSubmit}>
+      
+{items && items.length > 0 && <CourseList items={items} />}
+
+<Expander type="single" isCollapsible={true} defaultValue={["add"]}>
+<ExpanderItem title="Lisää uusi hakuvahti" value="add">
+      <h2>Lisää uusi hakuvahti</h2>
+
+      <div>
         <TextField label="Avainsanat" isRequired={true} onChange={handleKeywordChange} />
         <TextField label="Avainsanat, joita ei saa olla" onChange={handleNotKeywordChange} />
         <Button
           variation="primary"
           loadingText="Etsitään..."
+          onClick={handleAddSubscription}
           ariaLabel=""
           isLoading={!!isLoading}
           type="submit"
         >
-          Esikatsele kursseja
+          Lisää hakuvahti
         </Button>
-        <Button
-          variation="default"
-          loadingText="Etsitään..."
-          onClick={handleAddSubscription}
-          ariaLabel=""
-          isLoading={!!isLoading}
-        >
-          Ilmoita minulle uusista kursseista
-        </Button>
-      </form>
-
-      <Table
-  >
-    <TableHead>
-      <TableRow>
-        <TableCell as="th">Kurssi</TableCell>
-        <TableCell as="th">Ajankohta</TableCell>
-        <TableCell as="th">Paikka</TableCell>
-      </TableRow>
-    </TableHead>
-    <TableBody>
-    {items.map((value, index) => {
-          return <TableRow key={value.id}>
-          <TableCell>{value.name}</TableCell>
-          <TableCell>{value.timeinfo}</TableCell>
-          <TableCell>{value.location}</TableCell>
-        </TableRow>
-        })}
-      
-    </TableBody>
-  </Table>
+      </div>
+</ExpanderItem>
+      </Expander>
 
     </>
   );
